refactor: Move all util files to folder utils
