+++ src/index.ts
@@ -2,6 +2,9 @@ import SitumPlugin from "./sdk";
 
 // Definitions
 export * from "./sdk/types/index.d";
+
+// APIs
+export * from "./utils/requestPermission";
 export * from "./wayfinding";
 
 export default SitumPlugin;
