+++ src/utils/requestPermission.ts
@@ -10,68 +10,71 @@ import {
 const checkIOSPermissions = async () => {
   let granted = await request(PERMISSIONS.IOS.LOCATION_WHEN_IN_USE);
 
-  if (granted === RESULTS.GRANTED) {
-    console.info(
-      "Situm > permissionns > LOCATION_WHEN_IN_USE permission granted"
-    );
+  // Check if already denied
+  if (granted !== RESULTS.GRANTED) {
+    throw "Situm > permissionns > ACCESS_FINE_LOCATION denied";
+  }
 
-    //@ts-ignore
-    if (parseInt(Platform.Version, 10) > 12) {
-      granted = await request(PERMISSIONS.IOS.BLUETOOTH_PERIPHERAL);
+  console.debug(
+    "Situm > permissionns > LOCATION_WHEN_IN_USE permission granted"
+  );
 
-      if (granted === RESULTS.GRANTED) {
-        console.info(
-          "Situm > permissionns > BLUETOOTH_PERIPHERAL permission granted"
-        );
-        return true;
-      } else {
-        throw "Situm > permissionns > BLUETOOTH_PERIPHERAL permission not granted";
-      }
-    } else {
-      console.warn(
-        "Situm > permissionns > BLUETOOTH_PERIPHERAL permissions not required"
+  //@ts-ignore
+  if (parseInt(Platform.Version, 10) > 12) {
+    granted = await request(PERMISSIONS.IOS.BLUETOOTH_PERIPHERAL);
+
+    if (granted === RESULTS.GRANTED) {
+      console.debug(
+        "Situm > permissionns > BLUETOOTH_PERIPHERAL permission granted"
       );
       return true;
+    } else {
+      throw "Situm > permissionns > BLUETOOTH_PERIPHERAL permission not granted";
     }
   } else {
-    throw "Situm > permissionns > ACCESS_FINE_LOCATION denied";
+    console.debug(
+      "Situm > permissionns > BLUETOOTH_PERIPHERAL permissions not required"
+    );
+    return true;
   }
 };
 
 const checkAndroidPermissions = async () => {
   let granted;
+
   //@ts-ignore
-  if (Platform.Version > 30) {
-    console.log("Situm > permissionns > ANDROID VERSION > 30");
-    granted = await requestMultiple([
-      PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION,
-      PERMISSIONS.ANDROID.BLUETOOTH_CONNECT,
-      PERMISSIONS.ANDROID.BLUETOOTH_SCAN,
-    ]);
+  if (Platform.Version <= 30) {
+    console.debug("Situm > permissionns > ANDROID VERSION < 30");
 
-    if (
-      granted["android.permission.ACCESS_FINE_LOCATION"] === RESULTS.GRANTED &&
-      granted["android.permission.BLUETOOTH_CONNECT"] === RESULTS.GRANTED &&
-      granted["android.permission.BLUETOOTH_SCAN"] === RESULTS.GRANTED
-    ) {
-      console.info(
-        "Situm > permissionns > ACCESS_FINE_LOCATION, BLUETOOTH_CONNECT and BLUETOOTH_SCAN permissions granted"
-      );
-      return true;
-    }
-
-    throw "Situm > permissionns > ACCESS_FINE_LOCATION, BLUETOOTH_CONNECT and BLUETOOTH_SCAN permissions not granted";
-  } else {
-    console.info("ANDROID VERSION < 30");
     granted = await request(PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION);
 
     if (granted === RESULTS.GRANTED) {
-      console.info("Situm > permissionns > ACCESS_FINE_LOCATION granted");
+      console.debug("Situm > permissionns > ACCESS_FINE_LOCATION granted");
       return true;
     } else {
       throw "Situm > permissionns > ACCESS_FINE_LOCATION permission not granted";
     }
   }
+
+  console.debug("Situm > permissionns > ANDROID VERSION > 30");
+  granted = await requestMultiple([
+    PERMISSIONS.ANDROID.ACCESS_FINE_LOCATION,
+    PERMISSIONS.ANDROID.BLUETOOTH_CONNECT,
+    PERMISSIONS.ANDROID.BLUETOOTH_SCAN,
+  ]);
+
+  if (
+    granted["android.permission.ACCESS_FINE_LOCATION"] === RESULTS.GRANTED &&
+    granted["android.permission.BLUETOOTH_CONNECT"] === RESULTS.GRANTED &&
+    granted["android.permission.BLUETOOTH_SCAN"] === RESULTS.GRANTED
+  ) {
+    console.debug(
+      "Situm > permissionns > ACCESS_FINE_LOCATION, BLUETOOTH_CONNECT and BLUETOOTH_SCAN permissions granted"
+    );
+    return true;
+  }
+
+  throw "Situm > permissionns > ACCESS_FINE_LOCATION, BLUETOOTH_CONNECT or BLUETOOTH_SCAN permissions not granted";
 };
 
 const requestPermission = () =>
@@ -97,7 +100,7 @@ const requestPermission = () =>
         });
     }
 
-    reject(`Situm > permissionns > Platform ${Platform.OS} not supported`);
+    reject(`Situm > permissionns > Platform '${Platform.OS}' not supported`);
   });
 
 export default requestPermission;
